requirements.txt
pyusb==1.0.2
RPi.GPIO==0.7.0
